+++ libft/libft.h
@@ -6,41 +6,38 @@
 /*   By: jgil-cam <jgil-cam@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/24 13:18:39 by jgil-cam          #+#    #+#             */
-/*   Updated: 2022/05/06 16:52:45 by jgil-cam         ###   ########.fr       */
+/*   Updated: 2022/05/08 14:03:39 by jgil-cam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 
-#include <stdlib.h>
-#include <unistd.h>
-
 int				ft_isalpha(int c);
 int				ft_isdigit(int c);
 int				ft_isalnum(int c);
 int				ft_isascii(int c);
 int				ft_isprint(int c);
-size_t			ft_strlen(const char *s);
-void			*ft_memset(void *b, int c, size_t len);
-void			ft_bzero(void *s, size_t n);
-void			*ft_memcpy(void *dst, const void *src, size_t n);
-void			*ft_memmove(void *dst, const void *src, size_t len);
+unsigned int	ft_strlen(const char *s);
+void			*ft_memset(void *b, int c, unsigned int len);
+void			ft_bzero(void *s, unsigned int n);
+void			*ft_memcpy(void *dst, const void *src, unsigned int n);
+void			*ft_memmove(void *dst, const void *src, unsigned int len);
 unsigned int	ft_strlcpy(char *dst, const char *src, unsigned int dstsize);
-size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
+unsigned int	ft_strlcat(char *dst, const char *src, unsigned int dstsize);
 int				ft_toupper(int c);
 int				ft_tolower(int c);
 char			*ft_strchr(const char *s, int c);
 char			*ft_strrchr(const char *s, int c);
-int				ft_strncmp(const char *s1, const char *s2, size_t n);
-void			*ft_memchr(const void *s, int c, size_t n);
-int				ft_memcmp(const void *s1, const void *s2, size_t n);
+int				ft_strncmp(const char *s1, const char *s2, unsigned int n);
+void			*ft_memchr(const void *s, int c, unsigned int n);
+int				ft_memcmp(const void *s1, const void *s2, unsigned int n);
 char			*ft_strnstr(const char *haystack, const char *needle,
-					size_t len);
+					unsigned int len);
 int				ft_atoi(const char *str);
-void			*ft_calloc(size_t count, size_t size);
+void			*ft_calloc(unsigned long count, unsigned long size);
 char			*ft_strdup(const char *s1);
-char			*ft_substr(char const *s, unsigned int start, size_t len);
+char			*ft_substr(char const *s, unsigned int start, unsigned int len);
 char			*ft_strjoin(char const *s1, char const *s2);
 char			*ft_strtrim(char const *s1, char const *set);
 char			**ft_split(char const *s, char c);
